third_party/blink/web_tests/http/tests/uri/assign-path-with-leading-slash-expected.txt

The path assignment test redirected successfully. Checking path...

PASS location.pathname is '/uri/resources/assign-path-landing.html'
PASS successfullyParsed is true

TEST COMPLETE
